# chatgpt-plugin/.well-known/openapi.yaml
openapi: 3.0.1
info:
  title: Social Media Automation API
  description: API for uploading and sharing videos across social media platforms
  version: 'v1'
servers:
  - url: https://your-domain.com
paths:
  /upload:
    post:
      operationId: uploadVideo
      summary: Upload a video to social media platforms
      description: Process and upload a video to specified social media platforms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - videoUrl
                - platforms
              properties:
                videoUrl:
                  type: string
                  description: URL or content URI of the video to upload
                platforms:
                  type: array
                  items:
                    type: string
                    enum: [YOUTUBE_SHORTS, INSTAGRAM_REELS, TIKTOK]
                  description: List of platforms to upload to
                isCharity:
                  type: boolean
                  default: false
                  description: Whether to use charity account credentials
                context:
                  type: string
                  description: Context about the video for content generation
      responses:
        '200':
          description: Upload initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  uploadId:
                    type: string
                  deepLink:
                    type: string
                    description: Deep link to launch the Android app
  /status/{uploadId}:
    get:
      operationId: getUploadStatus
      summary: Get the status of a video upload
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current upload status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [PROCESSING, GENERATING_CONTENT, UPLOADING, COMPLETED, ERROR]
                  progress:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        platform:
                          type: string
                        stage:
                          type: string
                        progress:
                          type: number
                        message:
                          type: string